refactor: create component SocialLink
to use analytics for the links near the footer as well as the top bar.

refs:
issue #19
