+++ components/Socials.tsx
@@ -1,13 +1,12 @@
 import styled from "@emotion/styled";
-import { Link } from "@chakra-ui/core";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faFacebookSquare,
   faInstagram,
   faTwitterSquare,
 } from "@fortawesome/free-brands-svg-icons";
-import useStore from "../src/store";
 import React, { ReactElement } from "react";
+import SocialLink from "@/components/SocialLink";
 
 const StyledIcon = styled(FontAwesomeIcon)`
   width: 20px;
@@ -18,51 +17,29 @@ interface PropTypes {
 }
 
 export default function Socials(props: PropTypes): ReactElement {
-  const analytics: firebase.analytics.Analytics = useStore(
-    (state) => state.analytics
-  );
-
-  const clickSocial = (type: string) => {
-    if (analytics) {
-      analytics.logEvent(`click_social: ${type}`);
-    }
-  };
-
   return (
     <>
-      <Link
+      <SocialLink
         {...props}
+        eventType="facebook"
         href="https://www.facebook.com/cusecofficial/"
-        target="_blank"
-        rel="noopener noreferrer"
-        onClick={() => {
-          clickSocial("facebook");
-        }}
       >
         <StyledIcon icon={faFacebookSquare} />
-      </Link>
-      <Link
+      </SocialLink>
+      <SocialLink
         {...props}
+        eventType="twitter"
         href="https://twitter.com/cusec"
-        target="_blank"
-        rel="noopener noreferrer"
-        onClick={() => {
-          clickSocial("twitter");
-        }}
       >
         <StyledIcon icon={faTwitterSquare} />
-      </Link>
-      <Link
+      </SocialLink>
+      <SocialLink
         {...props}
+        eventType="instagram"
         href="https://www.instagram.com/cusecofficial/"
-        target="_blank"
-        rel="noopener noreferrer"
-        onClick={() => {
-          clickSocial("instagram");
-        }}
       >
         <StyledIcon icon={faInstagram} />
-      </Link>
+      </SocialLink>
     </>
   );
 }
