+++ components/sections/Contact.tsx
@@ -7,6 +7,7 @@ import FBIcon from "../svgs/facebook-coloured-icon.svg";
 import TwitterIcon from "../svgs/twitter-coloured-icon.svg";
 import SocialShareSVG from "../svgs/undraw-social-share.svg";
 import { TextStyled, TextStyledBold } from "@/components/StyledCore";
+import SocialLink from "../SocialLink";
 
 export default function Contact(): ReactElement {
   return (
@@ -31,16 +32,30 @@ export default function Contact(): ReactElement {
             justify={["center", "flex-start", "flex-start", "flex-start"]}
           >
             <Flex paddingRight={["0.15in", "0.15in", "0.2in", "0.25in"]}>
-              <FBIcon />
+              <SocialLink
+                eventType="facebook"
+                href="https://www.facebook.com/cusecofficial/"
+              >
+                <FBIcon />
+              </SocialLink>
             </Flex>
             <Flex paddingX={["0.15in", "0.15in", "0.2in", "0.25in"]}>
-              <TwitterIcon />
+              <SocialLink eventType="twitter" href="https://twitter.com/cusec">
+                <TwitterIcon />
+              </SocialLink>
             </Flex>
             <Flex paddingX={["0.15in", "0.15in", "0.2in", "0.25in"]}>
-              <IGIcon />
+              <SocialLink
+                eventType="instagram"
+                href="https://www.instagram.com/cusecofficial/"
+              >
+                <IGIcon />
+              </SocialLink>
             </Flex>
             <Flex paddingLeft={["0.15in", "0.15in", "0.2in", "0.25in"]}>
-              <EmailIcon />
+              <SocialLink eventType="email" href="mailto:info@cusec.net">
+                <EmailIcon />
+              </SocialLink>
             </Flex>
           </Flex>
           <Flex
